Role Based Authorization

+++ convex/files.ts
@@ -30,7 +30,7 @@ async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
         return null;
     }
 
-    const hasAccess = user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    const hasAccess = user.orgIds.some(item => item.orgId === orgId) || user.tokenIdentifier.includes(orgId);
 
     if (!hasAccess) {
         return null;
@@ -100,6 +100,12 @@ export const deleteFile = mutation({
             throw new ConvexError("No access to file");
         }
 
+        const isAdmin = access.user.orgIds.find(org => org.orgId === access.file.orgId)?.role === "admin";
+
+        if (!isAdmin) {
+            throw new ConvexError("You have no admin access to delete");
+        }
+
         await ctx.db.delete(args.fileId);
     },
 });
@@ -140,7 +146,7 @@ export const getAllFavorites = query({
                 q.eq('userId', hasAccess.user._id).eq("orgId", args.orgId)
             ).collect();
 
-            return favorites;
+        return favorites;
 
     },
 });
